pyproject.toml
[tool.poetry]
name = "devgrid"
version = "0.1.0"
description = ""
authors = ["Lucas Froguel <belasquelucas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
pytest = "^7.4.2"
faker = "^19.6.1"
pandas = "^2.1.0"
whitenoise = "^6.5.0"
djangorestframework-simplejwt = "^5.3.0"
django-health-check = "^3.17.0"

[tool.poetry.group.dev.dependencies]
python-decouple = "^3.8"
pymongo = "^4.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
